+++ extensions/vscode/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./exec";
+export * from "./prompt-for-target-directory";